requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
passlib[bcrypt]
python-jose
SQLAlchemy
psycopg2
fuzzywuzzy
python-Levenshtein
email_validator
